+++ webkit/plugins/ppapi/message_channel.cc
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2013 The Chromium Authors. All rights reserved.
+// Copyright (c) 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -93,7 +93,7 @@ bool PPVarToV8Value(PP_Var var, v8::Handle<v8::Value>* result) {
       // TODO(dmichael): We should consider caching the V8 string in the host-
       // side StringVar, so that we only have to convert/copy once if a
       // string is sent more than once.
-      *result = v8::String::New(value.c_str(), static_cast<int>(value.size()));
+      *result = v8::String::New(value.c_str(), value.size());
       break;
     }
     case PP_VARTYPE_ARRAY_BUFFER: {
